ninth commit

+++ src/main/java/org/example/datn_website_be/dto/request/CartDetailRequest.java
@@ -18,6 +18,6 @@ public class CartDetailRequest {
     private long idProduct;
     @NotNull(message = "Số lượng không được để trống")
     @Min(value = 1, message = "Số lượng phải lớn hơn hoặc bằng 1")
-    private Integer quantity;
+    private double quantity;
 
 }
