+++ src/main/java/org/example/datn_website_be/dto/request/CartRequest.java
@@ -15,5 +15,5 @@ public class CartRequest {
 
     @NotNull(message = "Số lượng của sản phẩm chi tiết là bắt buộc")
     @Positive(message = "Số lượng phải lớn hơn 0")
-    private Integer quantity;
+    private double quantity;
 }
