+++ src/main/java/org/example/datn_website_be/dto/response/PromotionResponse.java
@@ -21,7 +21,7 @@ public class PromotionResponse {
 
     private String name;
 
-    private Double value;
+    private Integer value;
 
     private Integer type;
