+++ src/main/java/org/example/datn_website_be/model/Voucher.java
@@ -37,7 +37,7 @@ public class Voucher extends BaseEntity {
     private String note;
 
     @Column
-    private Double value;
+    private Integer value;
 
     @Column
     private int quantity;
